backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
